utils-main/requirements.txt
pandas
pandas_ta
ta
ccxt
python-telegram-bot
Flask
scikit-learn
joblib